+++ tools/onnx-graphsurgeon/onnx_graphsurgeon/util/exception.py
@@ -15,4 +15,7 @@
 #
 
 class OnnxGraphSurgeonException(Exception):
+    """
+    An exception raised by ONNX-GraphSurgeon.
+    """
     pass